% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Socket.R
\name{close.Socket}
\alias{close.Socket}
\title{Close an R6 Socket with the standard R connection API.}
\usage{
\method{close}{Socket}(con, ...)
}
\arguments{
\item{con}{A \code{Socket} R6 object.}

\item{...}{Ignored.}
}
\description{
Close an R6 Socket with the standard R connection API.
}
