% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketr.R
\name{close.socketr_socket}
\alias{close.socketr_socket}
\title{Close a socketR handle with the standard R connection API.}
\usage{
\method{close}{socketr_socket}(con, ...)
}
\arguments{
\item{con}{A socketR socket handle.}

\item{...}{Ignored.}
}
\description{
Close a socketR handle with the standard R connection API.
}
