% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketr.R
\name{socket_get_option}
\alias{socket_get_option}
\title{Get a socket option.}
\usage{
socket_get_option(
  socket,
  level = "socket",
  option,
  type = c("int", "logical", "timeval", "linger", "raw"),
  size = 256L
)
}
\arguments{
\item{socket}{A socketR socket handle.}

\item{level}{Option level such as \code{"socket"} or \code{"tcp"}.}

\item{option}{Option name or native numeric constant.}

\item{type}{Return type: \code{"int"}, \code{"logical"}, \code{"timeval"}, \code{"linger"}, or \code{"raw"}.}

\item{size}{Maximum raw option size.}
}
\value{
The option value.
}
\description{
Get a socket option.
}
