% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketr.R
\name{socket_resolve}
\alias{socket_resolve}
\title{Resolve a hostname or address into socket endpoints.}
\usage{
socket_resolve(address, domain = c("inet", "inet6"), port = 0L)
}
\arguments{
\item{address}{A hostname or numeric IP address.}

\item{domain}{Address family, such as \code{"inet"} or \code{"inet6"}.}

\item{port}{Numeric service port between 0 and 65535.}
}
\value{
A list of resolved endpoint metadata.
}
\description{
Resolve a hostname or address into socket endpoints.
}
