% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socketr.R
\name{socket_lifecycle}
\alias{socket_lifecycle}
\alias{socket_close}
\alias{socket_is_open}
\alias{socket_shutdown}
\title{Manage a socket handle lifecycle.}
\usage{
socket_close(socket)

socket_is_open(socket)

socket_shutdown(socket, how = c("both", "read", "write"))
}
\arguments{
\item{socket}{A socketR socket handle.}

\item{how}{One of \code{"read"}, \code{"write"}, or \code{"both"}.}
}
\value{
Invisibly, \code{TRUE}.

A logical scalar.

Invisibly, the socket handle.
}
\description{
Manage a socket handle lifecycle.

Test whether a socket handle is open.

Shut down part of a full-duplex socket.
}
\examples{
\dontshow{if (.Platform$OS.type != "windows") withAutoprint(\{ # examplesIf}
s <- socket_create()
socket_close(s)
\dontshow{\}) # examplesIf}
}
